backend/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.7,<5.0
gunicorn>=20.0,<21.0
django-cors-headers>=3.7.0,<4.0